+++ awx/api/serializers/__init__.py
@@ -1,0 +1,3 @@
+from awx.api.serializers.base import *
+from awx.api.serializers.workflows import *
+from awx.api.serializers.unified_jobs import *
