+++ awx/api/serializers/base.py
@@ -1,0 +1,478 @@
+import copy
+
+from django.utils.encoding import force_str
+
+from rest_framework import serializers
+from rest_framework.relations import ManyRelatedField
+from rest_framework import fields
+
+from awx.main.utils import get_type_for_model
+
+
+class CopySerializer(serializers.Serializer):
+    name = serializers.CharField()
+
+    def validate(self, attrs):
+        name = attrs.get('name')
+        view = self.context.get('view', None)
+        obj = view.get_object()
+        if name == obj.name:
+            raise serializers.ValidationError(_('The original object is already named {}, a copy from' ' it cannot have the same name.'.format(name)))
+        return attrs
+
+
+class BaseSerializerMetaclass(serializers.SerializerMetaclass):
+    """
+    Custom metaclass to enable attribute inheritance from Meta objects on
+    serializer base classes.
+
+    Also allows for inheriting or updating field lists from base class(es):
+
+        class Meta:
+
+            # Inherit all fields from base class.
+            fields = ('*',)
+
+            # Inherit all fields from base class and add 'foo'.
+            fields = ('*', 'foo')
+
+            # Inherit all fields from base class except 'bar'.
+            fields = ('*', '-bar')
+
+            # Define fields as 'foo' and 'bar'; ignore base class fields.
+            fields = ('foo', 'bar')
+
+            # Extra field kwargs dicts are also merged from base classes.
+            extra_kwargs = {
+                'foo': {'required': True},
+                'bar': {'read_only': True},
+            }
+
+            # If a subclass were to define extra_kwargs as:
+            extra_kwargs = {
+                'foo': {'required': False, 'default': ''},
+                'bar': {'label': 'New Label for Bar'},
+            }
+
+            # The resulting value of extra_kwargs would be:
+            extra_kwargs = {
+                'foo': {'required': False, 'default': ''},
+                'bar': {'read_only': True, 'label': 'New Label for Bar'},
+            }
+
+            # Extra field kwargs cannot be removed in subclasses, only replaced.
+
+    """
+
+    @staticmethod
+    def _is_list_of_strings(x):
+        return isinstance(x, (list, tuple)) and all([isinstance(y, str) for y in x])
+
+    @staticmethod
+    def _is_extra_kwargs(x):
+        return isinstance(x, dict) and all([isinstance(k, str) and isinstance(v, dict) for k, v in x.items()])
+
+    @classmethod
+    def _update_meta(cls, base, meta, other=None):
+        for attr in dir(other):
+            if attr.startswith('_'):
+                continue
+            val = getattr(other, attr)
+            meta_val = getattr(meta, attr, None)
+            # Special handling for lists/tuples of strings (field names).
+            if cls._is_list_of_strings(val) and cls._is_list_of_strings(meta_val or []):
+                meta_val = meta_val or []
+                new_vals = []
+                except_vals = []
+                if base:  # Merge values from all bases.
+                    new_vals.extend([x for x in meta_val])
+                for v in val:
+                    if not base and v == '*':  # Inherit all values from previous base(es).
+                        new_vals.extend([x for x in meta_val])
+                    elif not base and v.startswith('-'):  # Except these values.
+                        except_vals.append(v[1:])
+                    else:
+                        new_vals.append(v)
+                val = []
+                for v in new_vals:
+                    if v not in except_vals and v not in val:
+                        val.append(v)
+                val = tuple(val)
+            # Merge extra_kwargs dicts from base classes.
+            elif cls._is_extra_kwargs(val) and cls._is_extra_kwargs(meta_val or {}):
+                meta_val = meta_val or {}
+                new_val = {}
+                if base:
+                    for k, v in meta_val.items():
+                        new_val[k] = copy.deepcopy(v)
+                for k, v in val.items():
+                    new_val.setdefault(k, {}).update(copy.deepcopy(v))
+                val = new_val
+            # Any other values are copied in case they are mutable objects.
+            else:
+                val = copy.deepcopy(val)
+            setattr(meta, attr, val)
+
+    def __new__(cls, name, bases, attrs):
+        meta = type('Meta', (object,), {})
+        for base in bases[::-1]:
+            cls._update_meta(base, meta, getattr(base, 'Meta', None))
+        cls._update_meta(None, meta, attrs.get('Meta', meta))
+        attrs['Meta'] = meta
+        return super(BaseSerializerMetaclass, cls).__new__(cls, name, bases, attrs)
+
+
+class BaseSerializer(serializers.ModelSerializer, metaclass=BaseSerializerMetaclass):
+    class Meta:
+        fields = ('id', 'type', 'url', 'related', 'summary_fields', 'created', 'modified', 'name', 'description')
+        summary_fields = ()
+        summarizable_fields = ()
+
+    # add the URL and related resources
+    type = serializers.SerializerMethodField()
+    url = serializers.SerializerMethodField()
+    related = serializers.SerializerMethodField('_get_related')
+    summary_fields = serializers.SerializerMethodField('_get_summary_fields')
+
+    # make certain fields read only
+    created = serializers.SerializerMethodField()
+    modified = serializers.SerializerMethodField()
+
+    def __init__(self, *args, **kwargs):
+        super(BaseSerializer, self).__init__(*args, **kwargs)
+        # The following lines fix the problem of being able to pass JSON dict into PrimaryKeyRelatedField.
+        data = kwargs.get('data', False)
+        if data:
+            for field_name, field_instance in self.fields.items():
+                if isinstance(field_instance, ManyRelatedField) and not field_instance.read_only:
+                    if isinstance(data.get(field_name, False), dict):
+                        raise serializers.ValidationError(_('Cannot use dictionary for %s' % field_name))
+
+    @property
+    def version(self):
+        return 2
+
+    def get_type(self, obj):
+        return get_type_for_model(self.Meta.model)
+
+    def get_types(self):
+        return [self.get_type(None)]
+
+    def get_type_choices(self):
+        type_name_map = {
+            'job': _('Playbook Run'),
+            'ad_hoc_command': _('Command'),
+            'project_update': _('SCM Update'),
+            'inventory_update': _('Inventory Sync'),
+            'system_job': _('Management Job'),
+            'workflow_job': _('Workflow Job'),
+            'workflow_job_template': _('Workflow Template'),
+            'job_template': _('Job Template'),
+        }
+        choices = []
+        for t in self.get_types():
+            name = _(type_name_map.get(t, force_str(get_model_for_type(t)._meta.verbose_name).title()))
+            choices.append((t, name))
+        return choices
+
+    def get_url(self, obj):
+        if obj is None or not hasattr(obj, 'get_absolute_url'):
+            return ''
+        elif isinstance(obj, User):
+            return self.reverse('api:user_detail', kwargs={'pk': obj.pk})
+        else:
+            return obj.get_absolute_url(request=self.context.get('request'))
+
+    def filter_field_metadata(self, fields, method):
+        """
+        Filter field metadata based on the request method.
+        This it intended to be extended by subclasses.
+        """
+        return fields
+
+    def _get_related(self, obj):
+        return {} if obj is None else self.get_related(obj)
+
+    def _generate_friendly_id(self, obj, node):
+        reset_counters()
+        return node.generate_named_url(obj)
+
+    def get_related(self, obj):
+        res = OrderedDict()
+        view = self.context.get('view', None)
+        if view and (hasattr(view, 'retrieve') or view.request.method == 'POST') and type(obj) in settings.NAMED_URL_GRAPH:
+            original_path = self.get_url(obj)
+            path_components = original_path.lstrip('/').rstrip('/').split('/')
+
+            friendly_id = self._generate_friendly_id(obj, settings.NAMED_URL_GRAPH[type(obj)])
+            path_components[-1] = friendly_id
+
+            new_path = '/' + '/'.join(path_components) + '/'
+            res['named_url'] = new_path
+        if getattr(obj, 'created_by', None):
+            res['created_by'] = self.reverse('api:user_detail', kwargs={'pk': obj.created_by.pk})
+        if getattr(obj, 'modified_by', None):
+            res['modified_by'] = self.reverse('api:user_detail', kwargs={'pk': obj.modified_by.pk})
+        return res
+
+    def _get_summary_fields(self, obj):
+        return {} if obj is None else self.get_summary_fields(obj)
+
+    def get_summary_fields(self, obj):
+        # Return values for certain fields on related objects, to simplify
+        # displaying lists of items without additional API requests.
+        summary_fields = OrderedDict()
+        for fk, related_fields in SUMMARIZABLE_FK_FIELDS.items():
+            try:
+                # A few special cases where we don't want to access the field
+                # because it results in additional queries.
+                if fk == 'job' and isinstance(obj, UnifiedJob):
+                    continue
+                if fk == 'project' and (isinstance(obj, InventorySource) or isinstance(obj, Project)):
+                    continue
+
+                try:
+                    fkval = getattr(obj, fk, None)
+                except ObjectDoesNotExist:
+                    continue
+                if fkval is None:
+                    continue
+                if fkval == obj:
+                    continue
+                summary_fields[fk] = OrderedDict()
+                for field in related_fields:
+                    fval = getattr(fkval, field, None)
+
+                    if fval is None and field == 'type':
+                        if isinstance(fkval, PolymorphicModel):
+                            fkval = fkval.get_real_instance()
+                        fval = get_type_for_model(fkval)
+                    elif fval is None and field == 'unified_job_type' and isinstance(fkval, UnifiedJobTemplate):
+                        fkval = fkval.get_real_instance()
+                        fval = get_type_for_model(fkval._get_unified_job_class())
+                    if fval is not None:
+                        summary_fields[fk][field] = fval
+            # Can be raised by the reverse accessor for a OneToOneField.
+            except ObjectDoesNotExist:
+                pass
+        if getattr(obj, 'created_by', None):
+            summary_fields['created_by'] = OrderedDict()
+            for field in SUMMARIZABLE_FK_FIELDS['user']:
+                summary_fields['created_by'][field] = getattr(obj.created_by, field)
+        if getattr(obj, 'modified_by', None):
+            summary_fields['modified_by'] = OrderedDict()
+            for field in SUMMARIZABLE_FK_FIELDS['user']:
+                summary_fields['modified_by'][field] = getattr(obj.modified_by, field)
+
+        # RBAC summary fields
+        roles = {}
+        for field in obj._meta.get_fields():
+            if type(field) is ImplicitRoleField:
+                roles[field.name] = role_summary_fields_generator(obj, field.name)
+        if len(roles) > 0:
+            summary_fields['object_roles'] = roles
+
+        # Advance display of RBAC capabilities
+        if hasattr(self, 'show_capabilities'):
+            user_capabilities = self._obj_capability_dict(obj)
+            if user_capabilities:
+                summary_fields['user_capabilities'] = user_capabilities
+
+        return summary_fields
+
+    def _obj_capability_dict(self, obj):
+        """
+        Returns the user_capabilities dictionary for a single item
+        If inside of a list view, it runs the prefetching algorithm for
+        the entire current page, saves it into context
+        """
+        view = self.context.get('view', None)
+        parent_obj = None
+        if view and hasattr(view, 'parent_model') and hasattr(view, 'get_parent_object'):
+            parent_obj = view.get_parent_object()
+        if view and view.request and view.request.user:
+            capabilities_cache = {}
+            # if serializer has parent, it is ListView, apply page capabilities prefetch
+            if self.parent and hasattr(self, 'capabilities_prefetch') and self.capabilities_prefetch:
+                qs = self.parent.instance
+                if 'capability_map' not in self.context:
+                    if hasattr(self, 'polymorphic_base'):
+                        model = self.polymorphic_base.Meta.model
+                        prefetch_list = self.polymorphic_base.capabilities_prefetch
+                    else:
+                        model = self.Meta.model
+                        prefetch_list = self.capabilities_prefetch
+                    self.context['capability_map'] = prefetch_page_capabilities(model, qs, prefetch_list, view.request.user)
+                if obj.id in self.context['capability_map']:
+                    capabilities_cache = self.context['capability_map'][obj.id]
+            return get_user_capabilities(
+                view.request.user, obj, method_list=self.show_capabilities, parent_obj=parent_obj, capabilities_cache=capabilities_cache
+            )
+        else:
+            # Contextual information to produce user_capabilities doesn't exist
+            return {}
+
+    def get_created(self, obj):
+        if obj is None:
+            return None
+        elif isinstance(obj, User):
+            return obj.date_joined
+        elif hasattr(obj, 'created'):
+            return obj.created
+        return None
+
+    def get_modified(self, obj):
+        if obj is None:
+            return None
+        elif isinstance(obj, User):
+            return obj.last_login  # Not actually exposed for User.
+        elif hasattr(obj, 'modified'):
+            return obj.modified
+        return None
+
+    def get_extra_kwargs(self):
+        extra_kwargs = super(BaseSerializer, self).get_extra_kwargs()
+        if self.instance:
+            read_only_on_update_fields = getattr(self.Meta, 'read_only_on_update_fields', tuple())
+            for field_name in read_only_on_update_fields:
+                kwargs = extra_kwargs.get(field_name, {})
+                kwargs['read_only'] = True
+                extra_kwargs[field_name] = kwargs
+        return extra_kwargs
+
+    def build_standard_field(self, field_name, model_field):
+        # DRF 3.3 serializers.py::build_standard_field() -> utils/field_mapping.py::get_field_kwargs() short circuits
+        # when a Model's editable field is set to False. The short circuit skips choice rendering.
+        #
+        # This logic is to force rendering choice's on an uneditable field.
+        # Note: Consider expanding this rendering for more than just choices fields
+        # Note: This logic works in conjuction with
+        if hasattr(model_field, 'choices') and model_field.choices:
+            was_editable = model_field.editable
+            model_field.editable = True
+
+        field_class, field_kwargs = super(BaseSerializer, self).build_standard_field(field_name, model_field)
+        if hasattr(model_field, 'choices') and model_field.choices:
+            model_field.editable = was_editable
+            if was_editable is False:
+                field_kwargs['read_only'] = True
+
+        # Pass model field default onto the serializer field if field is not read-only.
+        if model_field.has_default() and not field_kwargs.get('read_only', False):
+            field_kwargs['default'] = field_kwargs['initial'] = model_field.get_default()
+
+        # Enforce minimum value of 0 for PositiveIntegerFields.
+        if isinstance(model_field, (models.PositiveIntegerField, models.PositiveSmallIntegerField)) and 'choices' not in field_kwargs:
+            field_kwargs['min_value'] = 0
+
+        # Use custom boolean field that allows null and empty string as False values.
+        if isinstance(model_field, models.BooleanField) and not field_kwargs.get('read_only', False):
+            field_class = BooleanNullField
+
+        # Use custom char or choice field that coerces null to an empty string.
+        if isinstance(model_field, (models.CharField, models.TextField)) and not field_kwargs.get('read_only', False):
+            if 'choices' in field_kwargs:
+                field_class = ChoiceNullField
+            else:
+                field_class = CharNullField
+
+        # Update the message used for the unique validator to use capitalized
+        # verbose name; keeps unique message the same as with DRF 2.x.
+        opts = self.Meta.model._meta.concrete_model._meta
+        for validator in field_kwargs.get('validators', []):
+            if isinstance(validator, validators.UniqueValidator):
+                unique_error_message = model_field.error_messages.get('unique', None)
+                if unique_error_message:
+                    unique_error_message = unique_error_message % {'model_name': capfirst(opts.verbose_name), 'field_label': capfirst(model_field.verbose_name)}
+                    validator.message = unique_error_message
+
+        return field_class, field_kwargs
+
+    def build_relational_field(self, field_name, relation_info):
+        field_class, field_kwargs = super(BaseSerializer, self).build_relational_field(field_name, relation_info)
+        # Don't include choices for foreign key fields.
+        field_kwargs.pop('choices', None)
+        return field_class, field_kwargs
+
+    def get_unique_together_validators(self):
+        # Allow the model's full_clean method to handle the unique together validation.
+        return []
+
+    def run_validation(self, data=fields.empty):
+        try:
+            return super(BaseSerializer, self).run_validation(data)
+        except ValidationError as exc:
+            # Avoid bug? in DRF if exc.detail happens to be a list instead of a dict.
+            raise ValidationError(detail=serializers.as_serializer_error(exc))
+
+    def get_validation_exclusions(self, obj=None):
+        # Borrowed from DRF 2.x - return model fields that should be excluded
+        # from model validation.
+        cls = self.Meta.model
+        opts = cls._meta.concrete_model._meta
+        exclusions = [field.name for field in opts.fields]
+        for field_name, field in self.fields.items():
+            field_name = field.source or field_name
+            if field_name not in exclusions:
+                continue
+            if field.read_only:
+                continue
+            if isinstance(field, serializers.Serializer):
+                continue
+            exclusions.remove(field_name)
+        # The clean_ methods cannot be ran on many-to-many models
+        exclusions.extend([field.name for field in opts.many_to_many])
+        return exclusions
+
+    def validate(self, attrs):
+        attrs = super(BaseSerializer, self).validate(attrs)
+        try:
+            # Create/update a model instance and run its full_clean() method to
+            # do any validation implemented on the model class.
+            exclusions = self.get_validation_exclusions(self.instance)
+            obj = self.instance or self.Meta.model()
+            for k, v in attrs.items():
+                if k not in exclusions:
+                    setattr(obj, k, v)
+            obj.full_clean(exclude=exclusions)
+            # full_clean may modify values on the instance; copy those changes
+            # back to attrs so they are saved.
+            for k in attrs.keys():
+                if k not in exclusions:
+                    attrs[k] = getattr(obj, k)
+        except DjangoValidationError as exc:
+            # DjangoValidationError may contain a list or dict; normalize into a
+            # dict where the keys are the field name and the values are a list
+            # of error messages, then raise as a DRF ValidationError.  DRF would
+            # normally convert any DjangoValidationError to a non-field specific
+            # error message; here we preserve field-specific errors raised from
+            # the model's full_clean method.
+            d = exc.update_error_dict({})
+            for k, v in d.items():
+                v = v if isinstance(v, list) else [v]
+                v2 = []
+                for e in v:
+                    if isinstance(e, DjangoValidationError):
+                        v2.extend(list(e))
+                    elif isinstance(e, list):
+                        v2.extend(e)
+                    else:
+                        v2.append(e)
+                d[k] = list(map(force_str, v2))
+            raise ValidationError(d)
+        return attrs
+
+    def reverse(self, *args, **kwargs):
+        kwargs['request'] = self.context.get('request')
+        return reverse(*args, **kwargs)
+
+    @property
+    def is_detail_view(self):
+        if 'view' in self.context:
+            if 'pk' in self.context['view'].kwargs:
+                return True
+        return False
+
+
+class EmptySerializer(serializers.Serializer):
+    pass
