+++ awx/api/serializers/unified_jobs.py
@@ -1,0 +1,277 @@
+from rest_framework import serializers
+
+from awx.api.serializers.base import *
+
+from awx.main.models import UnifiedJob, UnifiedJobTemplate
+
+
+class UnifiedJobTemplateSerializer(BaseSerializer):
+    # As a base serializer, the capabilities prefetch is not used directly,
+    # instead they are derived from the Workflow Job Template Serializer and the Job Template Serializer, respectively.
+    capabilities_prefetch = []
+
+    class Meta:
+        model = UnifiedJobTemplate
+        fields = ('*', 'last_job_run', 'last_job_failed', 'next_job_run', 'status', 'execution_environment')
+
+    def get_related(self, obj):
+        res = super(UnifiedJobTemplateSerializer, self).get_related(obj)
+        if obj.current_job:
+            res['current_job'] = obj.current_job.get_absolute_url(request=self.context.get('request'))
+        if obj.last_job:
+            res['last_job'] = obj.last_job.get_absolute_url(request=self.context.get('request'))
+        if obj.next_schedule:
+            res['next_schedule'] = obj.next_schedule.get_absolute_url(request=self.context.get('request'))
+        if obj.execution_environment_id:
+            res['execution_environment'] = self.reverse('api:execution_environment_detail', kwargs={'pk': obj.execution_environment_id})
+        return res
+
+    def get_types(self):
+        if type(self) is UnifiedJobTemplateSerializer:
+            return ['project', 'inventory_source', 'job_template', 'system_job_template', 'workflow_job_template']
+        else:
+            return super(UnifiedJobTemplateSerializer, self).get_types()
+
+    def get_sub_serializer(self, obj):
+        serializer_class = None
+        if type(self) is UnifiedJobTemplateSerializer:
+            if isinstance(obj, Project):
+                serializer_class = ProjectSerializer
+            elif isinstance(obj, InventorySource):
+                serializer_class = InventorySourceSerializer
+            elif isinstance(obj, JobTemplate):
+                serializer_class = JobTemplateSerializer
+            elif isinstance(obj, SystemJobTemplate):
+                serializer_class = SystemJobTemplateSerializer
+            elif isinstance(obj, WorkflowJobTemplate):
+                serializer_class = WorkflowJobTemplateSerializer
+            elif isinstance(obj, WorkflowApprovalTemplate):
+                serializer_class = WorkflowApprovalTemplateSerializer
+        return serializer_class
+
+    def to_representation(self, obj):
+        serializer_class = self.get_sub_serializer(obj)
+        if serializer_class:
+            serializer = serializer_class(instance=obj, context=self.context)
+            # preserve links for list view
+            if self.parent:
+                serializer.parent = self.parent
+                serializer.polymorphic_base = self
+                # capabilities prefetch is only valid for these models
+                if isinstance(obj, (JobTemplate, WorkflowJobTemplate)):
+                    serializer.capabilities_prefetch = serializer_class.capabilities_prefetch
+                else:
+                    serializer.capabilities_prefetch = None
+            return serializer.to_representation(obj)
+        else:
+            return super(UnifiedJobTemplateSerializer, self).to_representation(obj)
+
+    def get_summary_fields(self, obj):
+        summary_fields = super().get_summary_fields(obj)
+
+        if self.is_detail_view:
+            resolved_ee = obj.resolve_execution_environment()
+            if resolved_ee is not None:
+                summary_fields['resolved_environment'] = {
+                    field: getattr(resolved_ee, field, None)
+                    for field in SUMMARIZABLE_FK_FIELDS['execution_environment']
+                    if getattr(resolved_ee, field, None) is not None
+                }
+
+        return summary_fields
+
+
+class UnifiedJobSerializer(BaseSerializer):
+    show_capabilities = ['start', 'delete']
+    event_processing_finished = serializers.BooleanField(
+        help_text=_('Indicates whether all of the events generated by this ' 'unified job have been saved to the database.'), read_only=True
+    )
+
+    class Meta:
+        model = UnifiedJob
+
+        fields = (
+            '*',
+            'unified_job_template',
+            'launch_type',
+            'status',
+            'execution_environment',
+            'failed',
+            'started',
+            'finished',
+            'canceled_on',
+            'elapsed',
+            'job_args',
+            'job_cwd',
+            'job_env',
+            'job_explanation',
+            'execution_node',
+            'controller_node',
+            'result_traceback',
+            'event_processing_finished',
+            'launched_by',
+            'work_unit_id',
+        )
+
+        extra_kwargs = {
+            'unified_job_template': {'source': 'unified_job_template_id', 'label': 'unified job template'},
+            'job_env': {'read_only': True, 'label': 'job_env'},
+        }
+
+    def get_types(self):
+        if type(self) is UnifiedJobSerializer:
+            return ['project_update', 'inventory_update', 'job', 'ad_hoc_command', 'system_job', 'workflow_job']
+        else:
+            return super(UnifiedJobSerializer, self).get_types()
+
+    def get_related(self, obj):
+        res = super(UnifiedJobSerializer, self).get_related(obj)
+        if obj.unified_job_template:
+            res['unified_job_template'] = obj.unified_job_template.get_absolute_url(request=self.context.get('request'))
+        if obj.schedule:
+            res['schedule'] = obj.schedule.get_absolute_url(request=self.context.get('request'))
+        if isinstance(obj, ProjectUpdate):
+            res['stdout'] = self.reverse('api:project_update_stdout', kwargs={'pk': obj.pk})
+        elif isinstance(obj, InventoryUpdate):
+            res['stdout'] = self.reverse('api:inventory_update_stdout', kwargs={'pk': obj.pk})
+        elif isinstance(obj, Job):
+            res['stdout'] = self.reverse('api:job_stdout', kwargs={'pk': obj.pk})
+        elif isinstance(obj, AdHocCommand):
+            res['stdout'] = self.reverse('api:ad_hoc_command_stdout', kwargs={'pk': obj.pk})
+        if obj.workflow_job_id:
+            res['source_workflow_job'] = self.reverse('api:workflow_job_detail', kwargs={'pk': obj.workflow_job_id})
+        if obj.execution_environment_id:
+            res['execution_environment'] = self.reverse('api:execution_environment_detail', kwargs={'pk': obj.execution_environment_id})
+        return res
+
+    def get_summary_fields(self, obj):
+        summary_fields = super(UnifiedJobSerializer, self).get_summary_fields(obj)
+        if obj.spawned_by_workflow:
+            summary_fields['source_workflow_job'] = {}
+            try:
+                summary_obj = obj.unified_job_node.workflow_job
+            except UnifiedJob.unified_job_node.RelatedObjectDoesNotExist:
+                return summary_fields
+
+            for field in SUMMARIZABLE_FK_FIELDS['job']:
+                val = getattr(summary_obj, field, None)
+                if val is not None:
+                    summary_fields['source_workflow_job'][field] = val
+
+        if self.is_detail_view:
+            ancestor = obj.ancestor_job
+            if ancestor != obj:
+                summary_fields['ancestor_job'] = {
+                    'id': ancestor.id,
+                    'name': ancestor.name,
+                    'type': get_type_for_model(ancestor),
+                    'url': ancestor.get_absolute_url(),
+                }
+
+        return summary_fields
+
+    def get_sub_serializer(self, obj):
+        serializer_class = None
+        if type(self) is UnifiedJobSerializer:
+            if isinstance(obj, ProjectUpdate):
+                serializer_class = ProjectUpdateSerializer
+            elif isinstance(obj, InventoryUpdate):
+                serializer_class = InventoryUpdateSerializer
+            elif isinstance(obj, Job):
+                serializer_class = JobSerializer
+            elif isinstance(obj, AdHocCommand):
+                serializer_class = AdHocCommandSerializer
+            elif isinstance(obj, SystemJob):
+                serializer_class = SystemJobSerializer
+            elif isinstance(obj, WorkflowJob):
+                serializer_class = WorkflowJobSerializer
+            elif isinstance(obj, WorkflowApproval):
+                serializer_class = WorkflowApprovalSerializer
+        return serializer_class
+
+    def to_representation(self, obj):
+        serializer_class = self.get_sub_serializer(obj)
+        if serializer_class:
+            serializer = serializer_class(instance=obj, context=self.context)
+            # preserve links for list view
+            if self.parent:
+                serializer.parent = self.parent
+                serializer.polymorphic_base = self
+                # TODO: restrict models for capabilities prefetch, when it is added
+            ret = serializer.to_representation(obj)
+        else:
+            ret = super(UnifiedJobSerializer, self).to_representation(obj)
+
+        if 'elapsed' in ret:
+            if obj and obj.pk and obj.started and not obj.finished:
+                td = now() - obj.started
+                ret['elapsed'] = (td.microseconds + (td.seconds + td.days * 24 * 3600) * 10**6) / (10**6 * 1.0)
+            ret['elapsed'] = float(ret['elapsed'])
+        # Because this string is saved in the db in the source language,
+        # it must be marked for translation after it is pulled from the db, not when set
+        ret['job_explanation'] = _(obj.job_explanation)
+        return ret
+
+    def get_launched_by(self, obj):
+        if obj is not None:
+            return obj.launched_by
+
+
+class UnifiedJobListSerializer(UnifiedJobSerializer):
+    class Meta:
+        fields = ('*', '-job_args', '-job_cwd', '-job_env', '-result_traceback', '-event_processing_finished')
+
+    def get_field_names(self, declared_fields, info):
+        field_names = super(UnifiedJobListSerializer, self).get_field_names(declared_fields, info)
+        # Meta multiple inheritance and -field_name options don't seem to be
+        # taking effect above, so remove the undesired fields here.
+        return tuple(x for x in field_names if x not in ('job_args', 'job_cwd', 'job_env', 'result_traceback', 'event_processing_finished'))
+
+    def get_types(self):
+        if type(self) is UnifiedJobListSerializer:
+            return ['project_update', 'inventory_update', 'job', 'ad_hoc_command', 'system_job', 'workflow_job']
+        else:
+            return super(UnifiedJobListSerializer, self).get_types()
+
+    def get_sub_serializer(self, obj):
+        serializer_class = None
+        if type(self) is UnifiedJobListSerializer:
+            if isinstance(obj, ProjectUpdate):
+                serializer_class = ProjectUpdateListSerializer
+            elif isinstance(obj, InventoryUpdate):
+                serializer_class = InventoryUpdateListSerializer
+            elif isinstance(obj, Job):
+                serializer_class = JobListSerializer
+            elif isinstance(obj, AdHocCommand):
+                serializer_class = AdHocCommandListSerializer
+            elif isinstance(obj, SystemJob):
+                serializer_class = SystemJobListSerializer
+            elif isinstance(obj, WorkflowJob):
+                serializer_class = WorkflowJobListSerializer
+            elif isinstance(obj, WorkflowApproval):
+                serializer_class = WorkflowApprovalListSerializer
+        return serializer_class
+
+    def to_representation(self, obj):
+        serializer_class = self.get_sub_serializer(obj)
+        if serializer_class:
+            serializer = serializer_class(instance=obj, context=self.context)
+            ret = serializer.to_representation(obj)
+        else:
+            ret = super(UnifiedJobListSerializer, self).to_representation(obj)
+        if 'elapsed' in ret:
+            ret['elapsed'] = float(ret['elapsed'])
+        return ret
+
+
+class UnifiedJobStdoutSerializer(UnifiedJobSerializer):
+    result_stdout = serializers.SerializerMethodField()
+
+    class Meta:
+        fields = ('result_stdout',)
+
+    def get_types(self):
+        if type(self) is UnifiedJobStdoutSerializer:
+            return ['project_update', 'inventory_update', 'job', 'ad_hoc_command', 'system_job']
+        else:
+            return super(UnifiedJobStdoutSerializer, self).get_types()
